+++ cpp/locality/NeighborBond.h
@@ -1,5 +1,10 @@
+// Copyright (c) 2010-2020 The Regents of the University of Michigan
+// This file is from the freud project, released under the BSD 3-Clause License.
+
 #ifndef NEIGHBOR_BOND_H
 #define NEIGHBOR_BOND_H
+
+#include "VectorMath.h"
 
 namespace freud { namespace locality {
 
@@ -9,20 +14,38 @@ namespace freud { namespace locality {
  *  class defines the less than operator according to distance, making it
  *  possible to sort.
  */
-struct NeighborBond
+class NeighborBond
 {
+public:
     // For now, id = query_point_idx and ref_id = point_idx (into the NeighborQuery).
     constexpr NeighborBond() = default;
 
-    constexpr NeighborBond(unsigned int query_point_idx, unsigned int point_idx, float d = 0, float w = 1)
-        : query_point_idx(query_point_idx), point_idx(point_idx), distance(d), weight(w)
+    /*
+     * This constructor should only be used in cases where the correct distance/vector
+     * pair is known before instantiating this object to save time and eliminate
+     * redundancy. Common cases include NeighbQuery objects which have to compute
+     * the distance to know if a neighbor pair fits the query arguments and NeighborList
+     * objects which have already computed the distances corresponding to their vectors.
+     * */
+    constexpr NeighborBond(unsigned int query_point_idx, unsigned int point_idx, float d, float w,
+                           const vec3<float>& v)
+        : query_point_idx(query_point_idx), point_idx(point_idx), distance(d), weight(w), vector(v)
+    {}
+
+    /**
+     * This constructor should be used for the majority of cases when instantiating
+     * a NeighborBond object.
+     */
+    NeighborBond(unsigned int query_point_idx, unsigned int point_idx, float w, const vec3<float>& v)
+        : query_point_idx(query_point_idx), point_idx(point_idx), distance(std::sqrt(dot(v, v))), weight(w),
+          vector(v)
     {}
 
     //! Equality checks both query_point_idx and distance.
     bool operator==(const NeighborBond& other) const
     {
         return (query_point_idx == other.query_point_idx) && (point_idx == other.point_idx)
-            && (distance == other.distance);
+            && (distance == other.distance) && (vector == other.vector);
     }
 
     //! Not equals checks inverse of equality.
@@ -87,10 +110,58 @@ struct NeighborBond
         return weight < n.weight;
     }
 
+    unsigned int getQueryPointIdx() const
+    {
+        return query_point_idx;
+    }
+
+    void setQueryPointIdx(unsigned int idx)
+    {
+        query_point_idx = idx;
+    }
+
+    unsigned int getPointIdx() const
+    {
+        return point_idx;
+    }
+
+    void setPointIdx(unsigned int new_pidx)
+    {
+        point_idx = new_pidx;
+    }
+
+    float getWeight() const
+    {
+        return weight;
+    }
+
+    void setWeight(float w)
+    {
+        weight = w;
+    }
+
+    const vec3<float>& getVector() const
+    {
+        return vector;
+    }
+
+    void setVector(const vec3<float>& v)
+    {
+        vector = v;
+        distance = std::sqrt(dot(v, v));
+    }
+
+    float getDistance() const
+    {
+        return distance;
+    }
+
+private:
     unsigned int query_point_idx {0}; //! The query point index.
     unsigned int point_idx {0};       //! The reference point index.
     float distance {0};               //! The distance between the points.
     float weight {0};                 //! The weight of this bond.
+    vec3<float> vector;               //! The directed vector from query point to reference point.
 };
 
 }; }; // end namespace freud::locality